requirements.txt
Flask==0.12.1
Flask-Cors==3.0.2
gunicorn==19.8.1
Pillow==4.0.0
